Added transcription to inquiry input
